+++ protocols/gossipsub/src/gossip_promises.rs
@@ -23,7 +23,6 @@ use crate::MessageId;
 use crate::ValidationError;
 use instant::Instant;
 use libp2p_identity::PeerId;
-use log::debug;
 use std::collections::HashMap;
 
 /// Tracks recently sent `IWANT` messages and checks if peers respond to them.
@@ -85,9 +84,10 @@ impl GossipPromises {
                 if *expires < now {
                     let count = result.entry(*peer_id).or_insert(0);
                     *count += 1;
-                    debug!(
-                        "[Penalty] The peer {} broke the promise to deliver message {} in time!",
-                        peer_id, msg
+                    tracing::debug!(
+                        peer=%peer_id,
+                        message=%msg,
+                        "[Penalty] The peer broke the promise to deliver message in time!"
                     );
                     false
                 } else {