+++ protocols/kad/src/behaviour/test.rs
@@ -321,7 +321,9 @@ fn query_iter() {
 
 #[test]
 fn unresponsive_not_returned_direct() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
+        .try_init();
     // Build one node. It contains fake addresses to non-existing nodes. We ask it to find a
     // random peer. We make sure that no fake address is returned.