+++ protocols/kad/tests/client_mode.rs
@@ -4,12 +4,15 @@ use libp2p_kad::store::MemoryStore;
 use libp2p_kad::{Behaviour, Config, Event, Mode};
 use libp2p_swarm::{Swarm, SwarmEvent};
 use libp2p_swarm_test::SwarmExt;
+use tracing_subscriber::EnvFilter;
 use Event::*;
 use MyBehaviourEvent::*;
 
 #[async_std::test]
 async fn server_gets_added_to_routing_table_by_client() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut client = Swarm::new_ephemeral(MyBehaviour::new);
     let mut server = Swarm::new_ephemeral(MyBehaviour::new);
@@ -32,7 +35,9 @@ async fn server_gets_added_to_routing_table_by_client() {
 
 #[async_std::test]
 async fn two_servers_add_each_other_to_routing_table() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut server1 = Swarm::new_ephemeral(MyBehaviour::new);
     let mut server2 = Swarm::new_ephemeral(MyBehaviour::new);
@@ -71,7 +76,9 @@ async fn two_servers_add_each_other_to_routing_table() {
 
 #[async_std::test]
 async fn adding_an_external_addresses_activates_server_mode_on_existing_connections() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut client = Swarm::new_ephemeral(MyBehaviour::new);
     let mut server = Swarm::new_ephemeral(MyBehaviour::new);
@@ -105,7 +112,9 @@ async fn adding_an_external_addresses_activates_server_mode_on_existing_connecti
 
 #[async_std::test]
 async fn set_client_to_server_mode() {
-    let _ = env_logger::try_init();
+    let _ = tracing_subscriber::fmt()
+        .with_env_filter(EnvFilter::from_default_env())
+        .try_init();
 
     let mut client = Swarm::new_ephemeral(MyBehaviour::new);
     client.behaviour_mut().kad.set_mode(Some(Mode::Client));