+++ protocols/relay/src/behaviour/handler.rs
@@ -339,7 +339,6 @@ pub struct Handler {
             <Self as ConnectionHandler>::OutboundProtocol,
             <Self as ConnectionHandler>::OutboundOpenInfo,
             <Self as ConnectionHandler>::ToBehaviour,
-            <Self as ConnectionHandler>::Error,
         >,
     >,
 
@@ -409,7 +408,7 @@ impl Handler {
             ))
             .is_err()
         {
-            log::warn!("Dropping inbound stream because we are at capacity")
+            tracing::warn!("Dropping inbound stream because we are at capacity")
         }
     }
 
@@ -432,7 +431,7 @@ impl Handler {
             )
             .is_err()
         {
-            log::warn!("Dropping outbound stream because we are at capacity")
+            tracing::warn!("Dropping outbound stream because we are at capacity")
         }
 
         self.active_connect_requests
@@ -482,7 +481,6 @@ type Futures<T> = FuturesUnordered<BoxFuture<'static, T>>;
 impl ConnectionHandler for Handler {
     type FromBehaviour = In;
     type ToBehaviour = Event;
-    type Error = void::Void;
     type InboundProtocol = ReadyUpgrade<StreamProtocol>;
     type InboundOpenInfo = ();
     type OutboundProtocol = ReadyUpgrade<StreamProtocol>;
@@ -505,7 +503,7 @@ impl ConnectionHandler for Handler {
                     ))
                     .is_some()
                 {
-                    log::warn!("Dropping existing deny/accept future in favor of new one.")
+                    tracing::warn!("Dropping existing deny/accept future in favor of new one")
                 }
             }
             In::DenyReservationReq {
@@ -519,7 +517,7 @@ impl ConnectionHandler for Handler {
                     ))
                     .is_some()
                 {
-                    log::warn!("Dropping existing deny/accept future in favor of new one.")
+                    tracing::warn!("Dropping existing deny/accept future in favor of new one")
                 }
             }
             In::NegotiateOutboundConnect {
@@ -588,16 +586,12 @@ impl ConnectionHandler for Handler {
         Instant::now().duration_since(idle_at) <= Duration::from_secs(10)
     }
 
+    #[tracing::instrument(level = "trace", name = "ConnectionHandler::poll", skip(self, cx))]
     fn poll(
         &mut self,
         cx: &mut Context<'_>,
     ) -> Poll<
-        ConnectionHandlerEvent<
-            Self::OutboundProtocol,
-            Self::OutboundOpenInfo,
-            Self::ToBehaviour,
-            Self::Error,
-        >,
+        ConnectionHandlerEvent<Self::OutboundProtocol, Self::OutboundOpenInfo, Self::ToBehaviour>,
     > {
         // Return queued events.
         if let Some(event) = self.queued_events.pop_front() {
@@ -651,11 +645,11 @@ impl ConnectionHandler for Handler {
                     ));
                 }
                 Poll::Ready(Err(e)) => {
-                    log::debug!("Inbound stream operation timed out: {e}");
+                    tracing::debug!("Inbound stream operation timed out: {e}");
                     continue;
                 }
                 Poll::Ready(Ok(Err(e))) => {
-                    log::debug!("Inbound stream operation failed: {e}");
+                    tracing::debug!("Inbound stream operation failed: {e}");
                     continue;
                 }
                 Poll::Pending => {
@@ -897,10 +891,7 @@ impl ConnectionHandler for Handler {
             ConnectionEvent::DialUpgradeError(dial_upgrade_error) => {
                 self.on_dial_upgrade_error(dial_upgrade_error);
             }
-            ConnectionEvent::AddressChange(_)
-            | ConnectionEvent::ListenUpgradeError(_)
-            | ConnectionEvent::LocalProtocolsChange(_)
-            | ConnectionEvent::RemoteProtocolsChange(_) => {}
+            _ => {}
         }
     }
 }