+++ swarm/src/connection.rs
@@ -192,7 +192,6 @@ where
                 ProtocolsChange::Added(ProtocolsAdded::from_set(&initial_protocols)),
             ));
         }
-
         Connection {
             muxing: muxer,
             handler,
@@ -235,10 +234,11 @@ where
 
     /// Polls the handler and the substream, forwarding events from the former to the latter and
     /// vice versa.
+    #[tracing::instrument(level = "debug", name = "Connection::poll", skip(self, cx))]
     pub(crate) fn poll(
         self: Pin<&mut Self>,
         cx: &mut Context<'_>,
-    ) -> Poll<Result<Event<THandler::ToBehaviour>, ConnectionError<THandler::Error>>> {
+    ) -> Poll<Result<Event<THandler::ToBehaviour>, ConnectionError>> {
         let Self {
             requested_substreams,
             muxing,
@@ -252,6 +252,7 @@ where
             remote_supported_protocols,
             idle_timeout,
             stream_counter,
+            ..
         } = self.get_mut();
 
         loop {
@@ -281,9 +282,6 @@ where
                 }
                 Poll::Ready(ConnectionHandlerEvent::NotifyBehaviour(event)) => {
                     return Poll::Ready(Ok(Event::Handler(event)));
-                }
-                Poll::Ready(ConnectionHandlerEvent::Close(err)) => {
-                    return Poll::Ready(Err(ConnectionError::Handler(err)));
                 }
                 Poll::Ready(ConnectionHandlerEvent::ReportRemoteProtocols(
                     ProtocolSupport::Added(protocols),
@@ -346,15 +344,15 @@ where
                     continue;
                 }
                 Poll::Ready(Some((_, Err(StreamUpgradeError::Io(e))))) => {
-                    log::debug!("failed to upgrade inbound stream: {e}");
+                    tracing::debug!("failed to upgrade inbound stream: {e}");
                     continue;
                 }
                 Poll::Ready(Some((_, Err(StreamUpgradeError::NegotiationFailed)))) => {
-                    log::debug!("no protocol could be agreed upon for inbound stream");
+                    tracing::debug!("no protocol could be agreed upon for inbound stream");
                     continue;
                 }
                 Poll::Ready(Some((_, Err(StreamUpgradeError::Timeout)))) => {
-                    log::debug!("inbound stream upgrade timed out");
+                    tracing::debug!("inbound stream upgrade timed out");
                     continue;
                 }
             }
@@ -451,9 +449,7 @@ where
     }
 
     #[cfg(test)]
-    fn poll_noop_waker(
-        &mut self,
-    ) -> Poll<Result<Event<THandler::ToBehaviour>, ConnectionError<THandler::Error>>> {
+    fn poll_noop_waker(&mut self) -> Poll<Result<Event<THandler::ToBehaviour>, ConnectionError>> {
         Pin::new(self).poll(&mut Context::from_waker(futures::task::noop_waker_ref()))
     }
 }
@@ -494,7 +490,7 @@ fn compute_new_shutdown(
 /// The [`Duration`] computed by the this function may not be the longest possible that we can add to `now` but it will work.
 fn checked_add_fraction(start: Instant, mut duration: Duration) -> Duration {
     while start.checked_add(duration).is_none() {
-        log::debug!("{start:?} + {duration:?} cannot be presented, halving duration");
+        tracing::debug!(start=?start, duration=?duration, "start + duration cannot be presented, halving duration");
 
         duration /= 2;
     }
@@ -541,7 +537,7 @@ impl<UserData, TOk, TErr> StreamUpgrade<UserData, TOk, TErr> {
     {
         let effective_version = match version_override {
             Some(version_override) if version_override != upgrade::Version::default() => {
-                log::debug!(
+                tracing::debug!(
                     "Substream upgrade protocol override: {:?} -> {:?}",
                     upgrade::Version::default(),
                     version_override
@@ -753,11 +749,14 @@ mod tests {
     use quickcheck::*;
     use std::sync::{Arc, Weak};
     use std::time::Instant;
+    use tracing_subscriber::EnvFilter;
     use void::Void;
 
     #[test]
     fn max_negotiating_inbound_streams() {
-        let _ = env_logger::try_init();
+        let _ = tracing_subscriber::fmt()
+            .with_env_filter(EnvFilter::from_default_env())
+            .try_init();
 
         fn prop(max_negotiating_inbound_streams: u8) {
             let max_negotiating_inbound_streams: usize = max_negotiating_inbound_streams.into();
@@ -924,7 +923,9 @@ mod tests {
 
     #[test]
     fn checked_add_fraction_can_add_u64_max() {
-        let _ = env_logger::try_init();
+        let _ = tracing_subscriber::fmt()
+            .with_env_filter(tracing_subscriber::EnvFilter::from_default_env())
+            .try_init();
         let start = Instant::now();
 
         let duration = checked_add_fraction(start, Duration::from_secs(u64::MAX));
@@ -934,7 +935,9 @@ mod tests {
 
     #[test]
     fn compute_new_shutdown_does_not_panic() {
-        let _ = env_logger::try_init();
+        let _ = tracing_subscriber::fmt()
+            .with_env_filter(EnvFilter::from_default_env())
+            .try_init();
 
         #[derive(Debug)]
         struct ArbitraryShutdown(Shutdown);
@@ -1104,7 +1107,7 @@ mod tests {
 
     #[derive(Default)]
     struct ConfigurableProtocolConnectionHandler {
-        events: Vec<ConnectionHandlerEvent<DeniedUpgrade, (), Void, Void>>,
+        events: Vec<ConnectionHandlerEvent<DeniedUpgrade, (), Void>>,
         active_protocols: HashSet<StreamProtocol>,
         local_added: Vec<Vec<StreamProtocol>>,
         local_removed: Vec<Vec<StreamProtocol>>,
@@ -1139,7 +1142,6 @@ mod tests {
     impl ConnectionHandler for MockConnectionHandler {
         type FromBehaviour = Void;
         type ToBehaviour = Void;
-        type Error = Void;
         type InboundProtocol = DeniedUpgrade;
         type OutboundProtocol = DeniedUpgrade;
         type InboundOpenInfo = ();
@@ -1195,7 +1197,6 @@ mod tests {
                 Self::OutboundProtocol,
                 Self::OutboundOpenInfo,
                 Self::ToBehaviour,
-                Self::Error,
             >,
         > {
             if self.outbound_requested {
@@ -1213,7 +1214,6 @@ mod tests {
     impl ConnectionHandler for ConfigurableProtocolConnectionHandler {
         type FromBehaviour = Void;
         type ToBehaviour = Void;
-        type Error = Void;
         type InboundProtocol = ManyProtocolsUpgrade;
         type OutboundProtocol = DeniedUpgrade;
         type InboundOpenInfo = ();
@@ -1272,7 +1272,6 @@ mod tests {
                 Self::OutboundProtocol,
                 Self::OutboundOpenInfo,
                 Self::ToBehaviour,
-                Self::Error,
             >,
         > {
             if let Some(event) = self.events.pop() {